LMS-1652 bugfix: create feature vector matrix taking the plate geometry into account, not the wells for which values have been provided
SVN: 17178

+++ screening/source/java/ch/systemsx/cisd/openbis/dss/generic/server/FeatureTableRow.java
@@ -35,7 +35,7 @@ public class FeatureTableRow
 
     private int columnIndex;
 
-    private double[] featureValues;
+    private float[] featureValues;
 
     public final String getDataSetCode()
     {
@@ -77,12 +77,12 @@ public class FeatureTableRow
         this.columnIndex = columnIndex;
     }
 
-    public final double[] getFeatureValues()
+    public final float[] getFeatureValues()
     {
         return featureValues;
     }
 
-    public final void setFeatureValues(double[] featureValues)
+    public final void setFeatureValues(float[] featureValues)
     {
         this.featureValues = featureValues;
     }
